Remove the description/workaround fields from Error definition
